Fix the SMWCLEUpdate motd behavior to restore old motd
Fix the SMWCLEUpdate motd behavior so that it restores the old /etc/motd
on the previous snapshot before rebooting into the new snapshot. This is
done so that the motd won't say the machine is updating if somebody
boots or chroots into an old snapshot.

Also make sure that the motd which includes the "UPDATE IN PROGRESS"
notice is written to the new snapshot while the update is still in
progress after the reboot.

Tested by putting the new run command in a task and executing it via fab
against an old Pinehurst VM. Checked that the task was successful
whether /etc/motd.bak existed or not.