IMPS-4710: Update SMWCLEUpdate motd for HipChat
Update the /etc/motd that is written out on the SMW by the SMWCLEUpdate
job, so that it refers to our CI HipChat room instead of the deprecated
imps IRC channel.

Test description:
Tested the update_motd task in the SMWCLEUpdate.py fabfile by setting
the JOB_NAME, BUILD_NUMBER, BUILD_URL, and REMOTESMW environment
variables and running the task against fireball-smw to verify that the
template is correct.