IMPS-5084: Add global cfgset update command to CleanInstallVirt
Add the command that is run during the documented fresh install process
for updating the global config set set to the CleanInstallVirt fabfile
that's used in both the CleanInstallVirt and RhineRedwoodFresh-* jenkins
jobs. Also add the validate command, but with --no-rules to skip the
rules since the global config set is not yet configured with actual data
at this point.

The point of doing this is to ensure that the documented fresh install
steps work in the order they are documented, specifically that 'cfgset
update -w /path/to/worksheets global' works before hardware discovery is
performed. This will help us catch bugs like IMPS-4497 in the future.

Although the CleanInstallVirt job's purpose is to create an SMW virtual
image for use on developer's personal machines, this additional testing
needs to be added to this job (as opposed to a separate job) because we
need to ensure that it works when run before hardware discovery which is
a part of this job.

The 'cfgset update [...] global' command can be disabled by setting the
environment variable UPDATE_GLOBAL_CONFIG_SET to false in the
environment.

Test Description:
Tested by running a CleanInstallVirt with this branch of the ci repo.