IMPS-6188: Store vagrant boxes and tmp in workspace
Set the VAGRANT_HOME environment variable to a subdirectory of the
WORKSPACE, so that when vagrant boxes are downloaded and stored, they
will only use space in the WORKSPACE of the jenkins job. If CLEANUP is
true on the jenkins job, this WORKSPACE directory will be removed at the
end of the jenkins job, so the extra box files and tmp files will not
take up valuable space in the jenkins home directory on our jenkins
servers.

Note that we seem to be using vagrant wrong here because we're using
what should really be the box url as the box name, and I think vagrant
is just smart enough to figure out where to find that box anyway. I
didn't try to clean this up because I consider that separate tech debt,
and it would be risky to fix because it might break other jobs that use
crayfab.

Test Description:
Ran ConfiguratorSystemTests using this branch and verified that no
ConfiguratorSystemTests vagrant box or tmp files were left in either the
workspace or under /var/lib/jenkins/.vagrant.d/.