IMPS-7190 Grab hacked ISO image
The fix in IMPS_6467 is needed to speed up our CI by 40+ minutes.
However, the SLEUpdate.iso, which contains this fix, is not released
on a very frequent schedule.  Thus, we might have had to wait a long
time to receive this fix in the official ISO.  Therefore, I created
a hacked version of the ISO containing this fix.  This commit now
imports that hacked ISO instead of the canonical one.