IMPS-7155 Set the currentBuild.result first; Work in the correct dir
Set the currentBuild.result to 'FAILURE' as the first thing in the
catch block.  That way if anything were to fail later in the block,
it does not prevent the currentBuild result from being correctly
labeled as a failure.

Issue the 'fab motd' command from within the correct directory where the
the fabfile.py is located.