IMPS-6550 Remove snapshots whose names conform to new format
The naming convention for snapshots was changed via IMPS_ , so that snapshots
could be retained or deleted on a per branch basis.  For example, if we want
to retain only the four most recent snapshots for a branch, we can do that
for the integration branch, the master branch, and a feature branch, or really
any branches we choose to designate in the RemoteHostCleanup Jenkins job's
parameters.

This mod creates remove_snap_new_name to remove snapshots that conform to
the new naming convention.

It parameterizes the following
SNAPS_TO_KEEP
IMAGES_TO_KEEP
CPIOS_TO_KEEP
These all become Jenkins job parameters.