IMPS-7418 Strip whitespace from branch names
Strip the whitespace from around the branch names that are used to
determine which branches to clean up.  This allows a comma separated
string of branch names to be used, and that string can contain whitespaces
between the commas.

Example;
"integration, master, feature"

Instead of requiring no spaces, which would be
"integration,master,feature"