IMPS-7234 Only change motd if CI tools downloaded
This mod creates a flag that indicates whether the Continuous
Integration (CI) tools were successfully downloaded or not.  If they
have been, then the jenkinsfile has the tools it needs to update the
Message Of The Day (motd) file on the SMW.  If not, then the jenkinsfile
should not attempt to update the motd file because the CI tools have not
been downloaded.  This mod prevents a spurious error about being unable
to modify the motd whenever a previous errors causes the job to abort
prior to the CI tools being downloaded.