CMS-2492 - remove HHMM from ci snapshot name as it is changing during the ci
We use the HHMM as part of CI_NAME, then we use CI_NAME all over the place.
It appears that CI_NAME is changing during the CI which is unexpected.  It
may be that those definitions are done for each FAB file, but for now we
shouldn't need it as we should have a unique build# anyways.