CMS-2448 - don't delete the snapshot when cleanup is checked
You almost never want to automatically delete a snapshot.  If the
CI succeeds, you typically want to use that as a working environment
till the next CI.  If the CI fails, you want to look at what went
wrong and try to fix it.  As a result, having the snapshot is
very valuable.

By removing the delete snapshot logic, this allows us to leave the
cleanup flag enabled in normal operation, which keeps the storage
use in check.