CMS-2803 - add additional flexibility to pe installation
Allow PE installation to be toggled on or off.

Allow building of the PE image to be toggled on or off.

Print hostname you are copying the image from to improve clarity.

Don't start to clean up and update the motd saying that we are finished if
we still need to boot the elogin.

Write the CI_NAME in the motd of the config set so you can see which ci is booted.

Unmount the mounted ISO in the cleanup first, in case something goes wrong with
subsequent actions.

Don't fail if the hwinv thingy isn't there

Remove extranious output.

Whitespace.

Corrected use of wrong variable