CMS-2944 - update the kernel during the ci
When a SLEupdate the user must manually set the kernel in the
snapshot so the CI should do the same, othewise the new kernel
will not be used.

Enable the bootmenu for each snapshot so the admin can easily switch
back and forth between snapshots.

Print the md5sum of the ISOs being used to improve tracability.

TESTING:

Tested with fireball-integration ci, overriding ci_branch.