CMS-2891 Integrate SMW eLogin into CI
This mod creates one or more eLogin nodes in the External State Daemon's (esd's)
node registry.  The nodes it creates come from the environmental variables passed
in by the Continuous Integration (CI) infrastructure.
It ensures that it deletes any pre-existing nodes in the registry that have the
same name, BMC_IP address, MGMT_MAC address, or MGMT_IP address as the nodes it
is trying to create.  It is illegal for there to be duplications.

It then PXE boots the nodes.

After a successful boot, it adds the kernel parameter STAGE=1 to the nodes'
entries in the registry and stages that to the nodes.

After a successful stage, it adds the kernel parameter STAGE=2 to the nodes'
entries in the registry.  Then it does a reboot --staged to the nodes.

It checks to see that the nodes have the STAGE=2 kernel parameter.  This
indicates that the proper stage was booted.  If the kernel parameter was
STAGED=1, then the previously staged image would have been booted, which was
not desirable.

Test Description:
- Ran through the program and saw that it booted and rebooted the nodes successfully.
- It handled an esd daemon that was not started by starting it.
- It handled deleting individual nodes whose BMC_IP, MGMT_IP, or MGMT_MAC
  conflicted with the node or nodes it was trying to create.
- It handled time-outs when it was waiting for nodes to boot and to stage and they
  failed to do so before the time-out expired.

In addition, this mod adds a test jenkinsfile directory.

This adds a 'test' directory under the jenkinsfiles directory.  This test
directory will hold a jenkinsfile that can be used for testing out jenkinsfile
changes that we may not want to test in the main jenkinsfile.  It is basically
a throw-away scratch file.

Finally, this mod implements a way to pass vairables
that are defined inside SMWCLEUpdate.py back to the jenkinsfile.
SMWCLEUpdate.py has a new task that creates a JSON file containing variables of
interest that the jenkinsfile can read.