CMS-3216 Change main CI to use eLogin variables
I repurposed the job variable ELOGIN_NODES to be the list of eLogin nodes that
we want to boot.  Thus, in the main CI job this variable, which was used to
decide whether to launch the eLogin job, was renamed to DO_ELOGIN_NODES.

I also re-ordered the parameters, so they were clustered with the non-eLogin-
specific parameters first and eLogin specific parameters second.

Test Description:
Launched a CI and saw it launch the job.