XCCS-170: CMS CI: SMWCLEUpdates needs to cleanup non-artifacts based on CI name and not snapshot name
This updates the pre_job_housekeeping task to use the most recent
snapshots which match the CI pattern to determine the list of artifacts
we wish to keep. Using this list of artifacs to keep, we can then
delete anything that that matches the CI_NAME pattern and is not
in the keep list.

We take the most recent snapshots to generate this list to keep and then
add any snapshot marked as cur, def, or ro.

Test Description:
Executed on my local machine and connected to fireball. Added a return
before any deletion occurs so I can inspect the output.

```
$ JOB_NAME='blake-test' BUILD_NUMBER=1 CI_BRANCH=integration PARTITION=p0 REMOTESMW=fireball-smw CLEANUP=true fab -f jobs/SMWCLEUpdate.py pre_job_housekeeping
```