XCCS-2098: Add wait between `enode delete` and `enode create`
Add a 5-second sleep between `enode delete` and `enode create` because
it was determined that some of the tasks queued by the delete of a
node can undo the work done by the tasks queued by the create of a node
if the create is run immediately after the delete.

Test Description:
On fireball-smw, tested manually that adding a sleep between the
commands results in correct dhcp configuration of the created node.