XCCS-2180: Moved update_firware after power_up_s0
xtzap --revsions expects that the system is powered up so it can talk to
all CCs and BCs. So we require the system to be up to print out the
revisions.

After running xtzap, it appears that it leaves the system in the same
state as xtcli power up s0 does.

This change moves update_firwmare right after power_up_s0.

Also fixed a number of trailing whitespace errors.