XCCS-2607: Add missing newline at end of motd
Add a newline that was missing at the end of the motd. I don't fully
understand why this combination of `-` in the jinja template works, but
it does.

Test Description:
Ran task against bubble-smw with updating=False and failed=False,
updating=True, and failed=True.  Also ran with INSTALL_BAREBONES set to
True and False in my environment.  Verified that the resulting /etc/motd
file always had a terminating newline by displaying it with `cat`.