XCCS-3090: CI_NAME should use MAIN_BRANCH to derive the name of the CI run for snapshots, images, etc.
I recently defaulted the CI_BRANCH to integration so that we can always
be sure that the CI jobs are using the correct CI infrastructure branch
to run the jobs in Jenkins. However, this messes with the naming scheme
(i.e. CI_NAME) since it derives the name from CI_BRANCH.

This changes CI_NAME to be derived with MAIN_BRANCH rather than
CI_BRANCH so that the resulting naming scheme is more what the user is
after.