XCCS-2709: Moved ci-test-elogin to run at the same time as bounce_and_boot_cle
This is so that ci-test-elogin and the eLogin tests as part of
ci-test-cle do not interfere with each other. extcheckout (ci-test-cle)
runs a suite of eLogin tests and ci-test-elogin turns off the elogin
node, causing the eLogin suite of tests to fail.

This change fires off ci-test-elogin at the same time CLE is booting.