XCCS-3350: Add `--no-pager` to `systemctl status esd`
Ensure that the `systemctl status esd` command does not try to use a
pager. Under certain circumstances that are not fully understood, it
will try to use a pager. This has been causing the ci job to hang when
it runs `systemctl status esd`. Using `--no-pager` should ensure that it
does not use a pager.

Test Description:
Tested that using `--no-pager` prevents a `systemctl status esd` command
from using a pager in a local test script that uses fabric to run that
command on fireball-smw. See comments in JIRA issue for more complete
description of that testing.