XCCS-3399: Update smoke test to ensure that the machine is booted
This adds a new fabric function, determine_system_state() that is now
part of the smoke test to ensure that the machine is booted, as
expected. We have an issue where xtbootsys does not return non-zero when
nodes time out.

This ensures looks for nodes that are not in 'ready' state and nodes
that have alerts. Failover nodes are excluded from this as they are
expected to be in 'standby' state. Disabled nodes are also exempt.

Test Description:
Ran the following test on fireball, bubble, pluto, orion, mars, and
flubber:

```
$ JOB_NAME=blake-test BUILD_NUMBER=4 REMOTESMW=mars-smw CI_BRANCH=integration BOOT_CLE=true fab -f src/jobs/SMWCLEUpdate.py smoke_test
```

The combo of the machine states helped me hit a large number of cases
for testing. flubber has failover nodes, for example.