XCCS-3258: Stop disabling firewall in CleanInstallVirt
Now that firewall configuration is fully functional in SLES 15, remove
the workaround that set `cray_firewall.enabled: false` in the global
config set.

Also remove a couple commands that disable the `SuSEfirewall2` and
`SuSEfirewalld_init` services. These no longer exist in SLES 15, and I'm
not even sure how these commands were passing.

Test Description:
Not sure how to test in the context of the civ jenkins job, but it works
to set `cray_firewall.enabled: true` in the global config set and run
`cray-ansible start` in the latest SMW VM from civ-integration-libvirt.