XCCS-3784: Remove DAL parameter from jobs
I noticed that there was some inconsistency in handling the DAL
nims group.  Specifically, we were setting up a parameter in the
jobs called "DAL", which was a comma-separated list of cnames
which would be assigned to the "dal" nims group during the CI.

Instead of using this parameter, it makes more sense
to specify which nodes are DAL nodes using the base cmap and using
node_groups in the base config set.  This is because even with
this DAL parameter and supporting code, you still have to modify
the cray_node_groups and possibly the cray_lustre_server and
cray_lustre_client configs to truly "assign" a node to be a DAL
node.

In the future it might be nice to have a more parameterized way of
assigning node functions and roles but as it stands I felt this
code was pretty broken.