XCCS-4056: Use GetSLEUpdate in pipeline jobs
The XCCS pipeline jobs (e.g. fireball-integration-pipeline) use
downstream Jenkins jobs to fetch or build media.  For example
the job which builds the CLE ISO is called Create-CLE-Media.
To fetch the SLEUpdate ISO, we were using the downstream job
ArchiveSLEUpdate.

ArchiveSLEUpdate itself has a downstream job called
GetSLEUpdate, which fetches the sle update ISO from a
known /cray/css/ location.  ArchiveSLEUpdate either calls its
downstream job GetSLEUpdate, or it calls a local packaging job,
formerly known as PackageSLEUpdates-local, but now does not exist.

The big problem is that ArchiveSLEUpdate is pretty useless and
confusing if it's just used right now as a passthrough to GetSLEUpdate.
But even if we were going to start packaging SLE updates locally again,
having an extra job in the middle is overkill.

Test Description:
Ran pluto-integration-pipeline with this revision for CI_BRANCH