XCCS-4056: Remove unused parameters
After removing ArchiveSLEUpdate, it is possible to remove
USE_PREBUILT_SLEUPDATE_ISO as a parameter to the jobs, which was
always set to true.  The entirety of the ArchiveSLEUpdate job was
to look at the USE_PREBUILT_SLEUPDATE_ISO environment variable and
if it was true, call GetSLEUpdate, and if it was false to call
PackageSLEUpdates-local (which does not exist anymore).

This commit also removes SLEUPDATE_INSTALLER_BRANCH, which looks
like a parameter that was used in PackageSLEUpdates-local.

I updated the documentation on the parameters as well.