XCCS-4056: Add GET_ISOS_FROM_EAP
Add a parameter to the pipeline jobs called GET_ISOS_FROM_EAP.  When
true, this will override the media creation behavior and instead fetch
the latest official master builds from CSS.

Also, this commit adds a change to use glob patterns to match the names
of the SMW and CLE ISOs so that their original names can be preserved
when copying them from EaP, but does not affect the behavior of
sleupdate ISOs.  This could be fixed in the future.

Finally, this commit enables the option on pluto-master-pipeline,
fireball-master-pipeline, and bubble-master-pipeline.

Test Description (for this and previous commit):
* First I ran dsl-seed locally to ensure the job-dsl changes were right
* Then I manually modified the job definition of pluto-master-pipeline
  * Remove USE_PREBUILT_SLEUPDATE_ISO
  * Remove SLEUPDATE_INSTALLER_BRANCH
  * Add GET_ISOS_FROM_EAP and set it to true
  * Set CI_BRANCH to this branch
* Ran pluto-master-pipeline