XCCS-4527: Split CLE testing jobs
This changes the ci-test-cle job definition from being a single
job whose parameters decide where it runs to having several
jobs, each of which is meant to run on a specific machine.

The advantage to this is that it will be easier to see at a glance
what is succeeding and what is failing.

There is no real difference between the "integration" and "master"
jobs except their names.  The integration pipeline jobs will call
the integration cle jobs, and the master pipeline jobs will call
the master cle jobs.  In fact, any of the jobs could be made to
run on any of the hosts given the right parameters.

Test Description:
Created jobs using dsl-seed, and then ran bubble-integration-cle
to ensure the job was functional.