XCCS-4542: Fix detection of UNSTABLE
This commit changes how the job is set to unstable, fixing a bug
where even successful jobs were being marked unstable.

Also, when detecting autoverify failures, add to the description
so that this will be included in the daily status e-mail.

Test Description:
I ran autoverify on bubble with and without a good config set,
resulting in green and yellow jobs respectively.