XCCS-4547: Add extcheckout failures to build description
The goal of this is so that extcheckout failing is reported in the
build description, so that it is viewable in the daily status email
as well as in Jenkins.  Currently we have a link to the results, but
not a simple statement saying whether or not it passed.

* Similar to how we grep the autoverify log, we will search the tap
  file.  However, this only updates the build description, and does
  not change the build's status, as that is already done by
  TapPublisher.
* Rename foundFailures variable
* Use HTML paragraph elements instead of break elements

Test Description:
I ran fireball-integration-cle with CI_BRANCH set to this branch
and then I checked the status of the job.