XCCS-2765: Rewrite alsorun steps and add DVS tests
This commit moves the alsoclient.py code from SMWCLEUpdate.py into
ci-test-cle.jenkinsfile, reducing the overall number of layers between
the Jenkins pipeline and OSTest.  Previously the alsorun client was
being run on the SMW via Fabric, for no real reason.  This changes
the behavior so that the alsorun client is run on the Jenkins server.
Because the alsorun client is a REST API client, there is no reason that
it should run on the SMW, and the tests will still login via the login
node as they did previously.

This commit also creates a parameterized function for alsoclient.py,
which is called twice in the pipeline, for extcheckout and DVS tests.
This commit adds a stage to the pipeline to call DVS tests.  It removes
the equivalent code from SMWCLEUpdate.py.

In this version of ci-test-cle.jenkinsfile, there are fewer
subdirectories in the workspace.  Instead of checking out the CI
repo to $WORKSPACE/$BUILD_ID/ci, we just check it out in $WORKSPACE.
This removes many needless dir() blocks and saves the overall
amount of disk space that the job needs.

The function updates the current build description with a message
saying whether the tests passed or failed.  The text itself is a
link to the results page.

This commit adds boolean parameters to job-dsl for testing DVS.  By
default, the fireball jobs will run DVS tests, but none of the other
ones will.  There is a duration parameter which is 1H30M.

Finally, this commit adds a Jenkins parameter for the e-mail recipient
(xccs-ci@cray.com), which was previously hard-coded.

Test Description:
To test the Jenkinsfile change, I added necessary parameters to
fireball-integration-cle, and then I ran fireball-integration-cle
with CI_BRANCH set to this branch.
I also tested bubble-integration-cle with a modified autoverify
configuration.
To test the job-dsl changes, I ran dsl-seed locally and inspected
the resulting jobs.