XCCS-4884: Do workaround for snaputil SuSE-release issue
This workaround hard-codes an absolute path to the snaputil command.
I checked that this path existed on all of fireball/pluto/bubble.

This workaround is used so that an older version of snaputil does
not attempt to read /etc/SuSE-release on an OS version that does
not have this file.

Review feedback: reworded comment for TODO

Test Description: ran a CI pipeline with CI_BRANCH set to this branch.