XCCS-5004: Separate xthwinv parameter from BOOT_CLE parameter
Previously, the BOOT_CLE parameter was used to decide if a fake
xthwinv script would be copied into the HSS bin directory.  But if
I want to test CI on a system with real hardware without booting
CLE, I need a way to keep the normal xthwinv command and still set
BOOT_CLE to False.  This commit adds an extra parameter so that
BOOT_CLE can be False but the xthwinv behavior is still the same.

Test Description:
* Ran bubble CI with BOOT_CLE set to False
* Added FAKE_XTHWINV parameter and set it to False