XCCS-3160: Re-enable staged boot test
Re-enable the test for `enode stage` functionality that runs in the
$MACHINE-$BRANCH-elogin jenkins jobs. Also clean up the test a bit. It
was doing a few unnecessary things. For one thing, it was running `enode
stage` twice to stage two different values for the same kernel
parameter. That adds time to the test, and I don't think it adds much
value.

It was also doing an `enode stage` followed right afterwards
by an `enode reboot --staged`. That would do the staging operation
twice since `enode reboot --staged` is basically like an `enode stage;
enode reboot --disk`. That might have had something to do with the
reboot failures we were seeing, but I'm not sure.

Test Description:
Ran fireball-integration-elogin against this branch and verified that it
was successful. Checked the value of the kernel parameter to make sure
it was correct on the node after it had booted.