XCCS-5067: Add warn_only to clear_sdb_mysql()
This changes the clear_sdb_mysql() method to not fail when a
Fabric run() returns a nonzero.  This is essential because in this
function we are using commands to test whether the sdb node is
reachable via ssh, but don't want the job to fail if it is not.

Review feedback:
Instead of wrapping the whole function using with
settings(warn_only=True), I only wrap the lines which I expect could
raise exceptions (sge).

Test Description:
Ran moon-patch\-6.0.up07-pipeline with CI_BRANCH set to this branch,
making sure that the sdb node was down before I started.