XCCS-5078: Allow post-build debugging
First, this changes the behavior of CLEANUP.  When False, the VM
should be left running.  This also changes the virt_wrapper.py
script to stop shutting down the VM in error cases.

Next, this commit changes the post-build steps of the VM jobs,
simplifying how artifacts are archived and making the workspace
more manageable when CLEANUP is not set to true.

This removes the needless archive/ subdirectory.  Previously
we were moving artifacts into a subdirectory called archive/,
and then taking things out of there and archiving them using
the Jenkins archiveArtifacts step.

This change makes it so that we first  archive the artifact and
then remove the .box and .iso files.  So, when CLEANUP is not
set to true, the workspace will still contain a Vagrantfile
and a Vagrant configuration, but the workspace will not grow
dramatically in size.

Test Description:
For the changes to InstallVirt.py and virt_wrapper.py:
* Ran an installvirt job with CI_BRANCH as this branch.  After
  running the job with CLEANUP==false, I navigated to the workspace
  directory and accessed the VM.
* Ran dsl-seed locally to ensure the job parameter descriptions were
  changed as expected
For the changes to the post-build steps:
* I made the changes to the job by running dsl-seed on xccs-ci
  and running the VM jobs, with a mixture of CLEANUP and CREATE_BOXFILE
  arguments