XCCS-5108: Change sdb node ssh slightly
Quotes are used when sending the "rm" command so that the glob is
expanded on the sdb node.  This is essentially the same as using
an escape character, but I think it is easier to understand.

Test Description: I ran this task using fab and checked directory
contents before and after.