XCCS-5108: Use sync in clear_sdb_mysql
This changes the clear_sdb_mysql task to issue a sync command on the
sdb node.  This is because the CI currently uses `xtcli halt` to shut
down the nodes prior to rebooting the SMW, which can lead to filesystem
changes not being written to disk.

This commit also adds some comments describing the bug XCCS_5112, which
tracks converting the CI to use a clean shutdown instead of halting.

Test Description: ran the task again and then halted/warm booted the
sdb node, to be sure that the files were being deleted.