XCCS-1389: Remove unnecessary parallel stage
This removes a parallel stage where it is not needed.  The
"Start test: elogin" stage simply launches a downstream job
and returns (the 'wait' parameter is set to 'false').  So doing
it in parallel with 'bounce_and_boot_cle' doesn't really save
any time compared to just running the tasks in order, which is
simpler to read and modify.