XCCS-5194: Remove unnecessary parallel block
This removes a parallel block around the stage which launches the
downstream CLE testing.  This is no longer needed because it is a
single task.  I think this originally came from when both CLE
and elogin testing were run downstream to the main pipeline with
"wait: true", which resulted in the main pipeline waiting for all
the testing to be done.  Now the jobs are just launched and the
main pipeline does not wait, so parallel blocks don't do anything.
And again because the CLE job is the only job which launches at this
point, the parallel block does nothing because it was only wrapping
a single stage.