XCCS-5194: Re-add WITH_NETROOT
This partially reverts changes from XCCS\-1389.  This re-adds
the parameter WITH_NETROOT and makes it possible to skip building
netroot images again.

Test Description: To test the job-dsl changes I ran dsl-seed locally
and ensured the resulting jobs were getting a parameter called
WITH_NETROOT.
To test the python and Jenkinsfile changes, I manually added
WITH_NETROOT to CI jobs and tested with it both enabled and disabled,
as well as with CI_BRANCH set to this branch.