XCCS-5231: Add additional "power_down_s0" stage
This adds an additional stage to the Jenkins pipeline.  After
rebooting the SMW, call `xtcli power down s0`.  This was previously
done before rebooting the SMW to the newly-installed snapshot.  This
does the same operation again, in case the snapshot is showing the
hardware state as anything but powered down.

Test Description:
I ran bubble-master-pipeline with CI_BRANCH set to this branch,
and I checked that firmware updates were happening.