XCCS-5301: Fix build_and_push_image_group for UP07 and earlier
The --nims-map argument does not exist on earlier versions of
imgbuilder.  This commit modifies the task to look at the
CI_NAME to figure out whether or not that argument is available.
If not, we temporarily set a cmap as active prior to running imgbuilder,
and then revert.

Also, this commit makes a few minor changes:
* Improve error handling and add CIError for cmap creation.
* Add better doc string for build_and_push_image_group.

Test Description:
I ran build_and_push_additional_images using fabric on moon-smw.