XCCS-4959: Add a PREBUILT_ELOGIN_IMAGES parameter
This changes the elogin jobs to allow a list of images instead
of image recipes.  This allows the elogin job to complete faster
if an elogin image was built as part of an upstream job, or just
if you don't need to build a new image.

This also changes the cleanup portion to only remove the image
if it was not a prebuilt image, because otherwise it's subject
to the usual cleanup.

Test Description: I ran moon-patch\-7.0.up00-elogin with
PREBUILT_ELOGIN_IMAGES manually set, and CI_BRANCH set to
this branch.  I also ran dsl-seed.