XCCS-4959: Pass elogin image to elogin job
This changes the way the pipeline calls the downstream elogin
job.  By using a regular expression on the imgbuilder output,
the pipeline gets the name of the elogin image that was built
by this job and set PREBUILT_ELOGIN_IMAGES to it.

Test Description:
To test this change as well as the previous change, I added the
PREBUILT_ELOGIN_IMAGES parameter to the elogin job, and then I
ran the main pipeline.  I set CI_BRANCH for this job and for the
elogin job to this branch.