XCCS-5355: Ensure postgres is running prior to xtboune
This change adds a new task that uses `systemctl` to ensure that
postgres is running prior to attempting an `xtbounce`.  This is
only needed in rare cases when postgresql.service has failed to
start after an SMW reboot.  This assumes that pmdb_util has already
correctly configured the postgres database.

Test Description: I ran this task as a stand-alone on bubble-smw
with and without postgres running, ensuring that when it was running
the task did nothing, and when it was not running that the task
started postgres.  I also ran bubble-integration-pipeline with
CI_BRANCH overridden to this branch.