XCCS-5535: Add pushing of diags images to CI
This defines two new parameters for the pipeline jobs:
* DIAGS_X86_IMAGE
* DIAGS_ARM IMAGE

For each of the pipeline jobs, a stage is added to push the
images with the given names.  For pluto/fireball/bubble jobs, the
default is used, but for patch branch jobs, the image names
are overridden.

A task is added to the fabfile to run the `image push` command.
For now it is set up to not raise an exception on failure but instead
log a warning.

Test Description:
To test the job-dsl changes, I ran dsl-seed locally and inspected
the resulting jobs.  To test the Jenkinsfile and fabfile changes,
I manually modified bubble-integration-pipeline to add DIAGS_X86_IMAGE
and DIAGS_ARM_IMAGE parameters, then set CI_BRANCH to this branch.
I ran the job and ensured it was pushing the diags image as expected.