XCCS-5765: Convert virt jobs into a pipeline
This changes VirtBuilder to use a pipeline.  This is just the
initial conversion to pipeline form.  There are many improvements
that can be made after this:
* split out stages into individual steps
* reduce duplicated code between Jenkinsfiles

Test Description: ran dsl-seed and several resulting jobs
with CI_BRANCH set to this branch, with different combinations
of parameters.  Note: when testing, the jobs were all prefixed
vmp-* rather than vm-*