XCCS-6055: Query REST API for build dates
A recent update to Jenkins has stopped providing a "published"
date via the RSS feed that the daily status job uses.  If the
date is missing, this commit makes the script use the Jenkins
REST API to get that information instead.

It adds a JenkinsJob class that will hopefully make development
against the REST API easier in the future as well.

Test Description:
I ran the daily_status job to send an e-mail to myself, then
I ran the Jenkins job with CI_BRANCH set to this branch.