XCCS-6109: Add a 10 minute timeout to power_down_s0
Nesting the 10 minute timeout within the retry block means that
there will be a total of 5 retries, and each iteration will be
capped at 10 minutes.  This is quite a bit more time than the
operation needs, as it usually takes around 1 minute to return.

Test Description: I ran pluto-integration-pipeline with CI_BRANCH
set to this branch.