XCCS-6169: Add xtdiscover with bounce to CI
This changes the CI process to add an optional xtdiscover stage,
which currently will only be enabled for flubber-master-pipeline.
Therefore a RUN_XTDISCOVER parameter is added to control whether
a job will include this step.

This is accomplished using a standalone script, autoxtdiscover.py
which is adapted from autosmwupdate, but modified to use Fabric
to work over SSH.  It is a departure from the way most of the CI
scripts work in that it provides input to the prompts using pexpect.

This commit also factors some bits RPM installation code into a function
called install_bits_rpms().

Because the discover includes a bounce it partially addresses the
issue of xtzap flashing the wrong bios on CLX nodes because it
leaves the nodes powered on.

Test Description: I ran flubber-master-pipeline with RUN_XTDISCOVER
added and CI_BRANCH set to this branch.