XCCS-6407: Make SUPPORTING_ISOS a parameter
This changes the InstallVirt.py script from using the RELEASE_DIST
parameter to figure out what ISOs to download to making the ISOs
possible to override at run time.  This is overall easier to maintain
because it is more transparent, and also makes RELEASE_DIST not as
overloaded of a parameter.

Test Description: added SUPPORTING_ISOS to an installvirt job and
then ran it with CI_BRANCH set to this branch