XCCS-5904: Stop storing ISOs in /tmp
This changes the CI job to download ISOs to /var/adm/cray/release
instead of /tmp, so that keeping ISOs from multiple CI branches
does not have a chance of filling up /tmp.  There is still a cleanup
policy that keeps the filesystem from filling up, but even with
cleanup, /tmp was not big enough.

Test Description: ran fireball-integration-pipeline with CI_BRANCH
set to this branch.