XCCS-6385: Validation should use SMWinstall, no --force
This changes all CI jobs to use SMWinstall instead of install.py.  The
SMWinstall command is a symlink to install.py, so the change is mostly
about being in line with the official install procedure.

This also changes the job-dsl configuration so that FORCEUPDATE is not
used in flubber validation.  We need to be validating that the installer
works on every update without using --force.

Test Description:
* To test the InstallVirt and SMWCLEUpdate changes, I ran CI jobs
  with CI_BRANCH set to this branch.
* To test the job-dsl change, I ran dsl-seed locally and inspected
  the new default value for FORCEUPDATE.