XCCS-5317: Fix date issue and rework get_job_vars()
This commit changes the CI_NAME variable so that it is a global
variable to the Jenkins pipeline, where previously it was global
to the python fabfile.

CI_NAME is a a job-specific identifier to name config sets,
snapshots, cmaps and images associated with a particular run of a
job.  The naming convention is:

"ci-<partition>-<branch>-<buildNumber>-<date>"

Because CI_NAME was previously a global variable in SMWCLEUpdate.py
(the fabfile), it would be re-evaluated with every call to "fab"
in the Jenkins pipeline (i.e. every stage).  If midnight passed
during a CI job, then CI_NAME would change, and the job would not
be able to refer to its images, config sets, snapshot, and cmap.
The variables are now passed to SMWCLEUpdate.py via the environment.

This commit also removes the get_job_vars() task and global job_vars
dictionary because they are no longer needed.  Finally, a needless
declaration of elogin job parameters was removed.

Finally this commit updates the copyright statement in the Jenkinsfile,
and removes an unused import from the fabfile.

Test Description: I ran the moon patch branch CI job with
CI_BRANCH set to this branch.