XCCS-7401: Fix reboot function
Fix the reboot function.  We should make sure that curr_uptime is less than
mark.  When doing a reboot, sshd might remain running for a while before the
machine is shutdown.  In this case, it would appear that the SMW is rebooted
when it is not, and cause a failure.