XCCS-8863: Move patch directory creation
Move the patch directory creation to prior to booting the nodes.
Moon (and only moon) uses the Configure_deliveral_patch_directories.groovy
script.  A lot of times an arm node will fail on moon due to known issues.

We want to copy the patch isos regardless of mainly the smoke test.