XCCS-8995: Autoverify/elogin testing
Since autoverify is testing the elogin, and the autoverify tests are
within the CLE testing, wait for the elogin testing to complete before
starting the CLE tests.

Run the remaining main pipeline jobs in parallel with the elogin tests
to save some time, but wait for the elogin tests prior to starting the
CLE tests.