feature/SCICD-65 Adjust for non-specific cos version in configuration
The COS version in the CI job has been shortened to just the release and no longer includes the full date string.  Modify the merge process to look for any matching COS releases in the system configuration and only pull in the latest dated string.

Also remove some redundancy in the integration branch name