SCICD-113: add-a-hello-world-test-to-run-on-the-compute-nodes
This change runs "hostname" on all the "idle" slurm nodes.
If the set of nodes in 'idle' state matches the hostnames
returned in the srun output, the step succeedes.
If the node sets don't match, the test fails. If
the 'idle' node count doesn't match the (sat reported) 'Ready'
node count,the step will print a warning.