SCICD-149 Add failure wrappers to the installer
Updated connection.sudo to throw exceptions on failure like it used to,
and now trap errors at the top level to cause the script to abort.
Output from the connection.sudo will be dumped to the debug log.
Updated the PE unmount process to run a remote script rather than
blindly running ssh commands and just not caring if they failed.