SCICD-205 Remove run_command in favor of using connection.sudo
Remove all the various instances of run_command in the InstallerUtils
and replace it with the standard command running interface.  Update the
command interface to allow dryrun to be overridden for functions that
need to be executed for data gathering for basic functionality.

Move the custom exception format to vars.py with the rest of the custom
exception formats.

Update get_git calls to pass in the connection object as it isn't
available to the utils by default.